feat: mainBox 디자인 수정, Swiper 분리
